Alex

+++ admin.py
@@ -9,21 +9,27 @@ from utils import build_reply_markup
 
 def admin_panel_processing(message):
     if message.text == constants.ADD_ADMIN:
+        markup = build_reply_markup([constants.CANCEL])
+
         msg = bot.send_message(
-            message.chat.id, "Введіть ім'я нового адміна:"
-        )
+            message.chat.id, "Введіть ім'я нового адміна:",
+            reply_markup=markup)
         bot.register_next_step_handler(msg, get_admin_id)
 
     elif message.text == constants.REMOVE_ADMIN:
+        markup = build_reply_markup([constants.CANCEL])
+
         msg = bot.send_message(
-            message.chat.id, "Введіть ID адміна, якого бажаєте видалити:"
-        )
+            message.chat.id, "Введіть ID адміна, якого бажаєте видалити:",
+            reply_markup=markup)
         bot.register_next_step_handler(msg, del_admin_from_db)
 
     elif message.text == constants.ADD_COLLEGE:
+        markup = build_reply_markup([constants.CANCEL])
+
         msg = bot.send_message(
-            message.chat.id, "Введіть коледж, який бажаєте додати:"
-        )
+            message.chat.id, "Введіть коледж, який бажаєте додати:",
+            reply_markup=markup)
         bot.register_next_step_handler(msg, process_college)
 
     elif message.text == constants.SEND_MESSAGE:
@@ -45,9 +51,21 @@ def admin_panel_processing(message):
 
 
 def get_admin_id(message):
-    name = message.text
-    msg = bot.send_message(message.chat.id, "Введіть ID нового адміна:")
-    bot.register_next_step_handler(msg, add_admin_to_db, name=name)
+    if message.text == constants.CANCEL:
+        if db.is_admin(message.from_user.id):
+            markup = build_reply_markup(
+                constants.ADMIN_PANEL_BUTTONS, one_time_keyboard=False)
+
+            msg = bot.send_message(
+                message.chat.id, "Ви знаходитесь в адмін панелі!",
+                reply_markup=markup)
+
+            bot.register_next_step_handler(msg, admin_panel_processing)
+
+    else:
+        name = message.text
+        msg = bot.send_message(message.chat.id, "Введіть ID нового адміна:")
+        bot.register_next_step_handler(msg, add_admin_to_db, name=name)
 
 
 def add_admin_to_db(message, name: str):
@@ -55,28 +73,58 @@ def add_admin_to_db(message, name: str):
     db.add_admin_to_db(telegram_id, name)
     bot.send_message(message.chat.id, "Адміна додано")
 
+    bot.register_next_step_handler(message, admin_panel_processing)
+
 
 def del_admin_from_db(message):
-    admin_id = message.text
-    db.remove_admin_from_db(admin_id)
-    bot.send_message(message.chat.id, "Адміна видалено")
+    if message.text == constants.CANCEL:
+        if db.is_admin(message.from_user.id):
+            markup = build_reply_markup(
+                constants.ADMIN_PANEL_BUTTONS, one_time_keyboard=False)
+
+            msg = bot.send_message(
+                message.chat.id, "Ви знаходитесь в адмін панелі!",
+                reply_markup=markup)
+
+            bot.register_next_step_handler(msg, admin_panel_processing)
+
+    else:
+        admin_id = message.text
+        db.remove_admin_from_db(admin_id)
+        bot.send_message(message.chat.id, "Адміна видалено")
+
+        bot.register_next_step_handler(message, admin_panel_processing)
 
 
 def process_college(message):
-    name = message.text
+    if message.text == constants.CANCEL:
+        if db.is_admin(message.from_user.id):
+            markup = build_reply_markup(
+                constants.ADMIN_PANEL_BUTTONS, one_time_keyboard=False)
 
-    markup = build_reply_markup(constants.LOCATIONS)
+            msg = bot.send_message(
+                message.chat.id, "Ви знаходитесь в адмін панелі!",
+                reply_markup=markup)
 
-    msg = bot.send_message(
-        message.chat.id, "Оберіть локацію коледжу:", reply_markup=markup
-    )
-    bot.register_next_step_handler(msg, add_college_to_db, name=name)
+            bot.register_next_step_handler(msg, admin_panel_processing)
+
+    else:
+        name = message.text
+
+        markup = build_reply_markup(constants.LOCATIONS)
+
+        msg = bot.send_message(
+            message.chat.id, "Оберіть локацію коледжу:", reply_markup=markup
+        )
+        bot.register_next_step_handler(msg, add_college_to_db, name=name)
 
 
 def add_college_to_db(message, name: str):
     college_location = message.text
     db.add_college_to_db(college_location, name)
     bot.send_message(message.chat.id, "Коледж додано")
+
+    bot.register_next_step_handler(message, admin_panel_processing)
 
 
 def admin_send_message(message):
@@ -91,6 +139,8 @@ def admin_send_message(message):
 
         bot.reply_to(message, 'Запити на оновлення даних надіслано')
 
+        bot.register_next_step_handler(message, admin_panel_processing)
+
     elif message.text == constants.CUSTOM_MESSAGE:
         msg = bot.send_message(
             message.chat.id,
@@ -98,12 +148,25 @@ def admin_send_message(message):
         bot.register_next_step_handler(
             msg, admin_send_custom_message_processing)
 
+    elif message.text == constants.CANCEL:
+        if db.is_admin(message.from_user.id):
+            markup = build_reply_markup(
+                constants.ADMIN_PANEL_BUTTONS, one_time_keyboard=False)
+
+            msg = bot.send_message(
+                message.chat.id, "Ви знаходитесь в адмін панелі!",
+                reply_markup=markup)
+
+            bot.register_next_step_handler(msg, admin_panel_processing)
+
 
 def admin_send_custom_message_processing(message):
     for user in db.get_telegram_id_list():
         bot.send_message(user, message.text)
 
     bot.reply_to(message, 'Повідомллення надіслані')
+
+    bot.register_next_step_handler(message, admin_panel_processing)
 
 
 def admin_send_files_csv(message: types.Message, type: str):
@@ -121,3 +184,5 @@ def admin_send_files_csv(message: types.Message, type: str):
 
     if os.path.exists(filename):
         os.remove(filename)
+
+    bot.register_next_step_handler(message, admin_panel_processing)
